+++ lib/brando/sites/global_set.ex
@@ -47,7 +47,7 @@ defmodule Brando.Sites.GlobalSet do
         fieldset do
           inputs_for :vars do
             label t("Globals")
-            component BrandoAdmin.Components.Form.Input.Vars
+            component :vars
           end
         end
       end